+++ src/components/SidePanel/SidePanel.tsx
@@ -1,10 +1,7 @@
-import { useState, useEffect } from 'react'
-import { AppShell, Button, Group, Stack, Title, Menu } from '@mantine/core'
-import { IconAffiliate, IconChevronRight, IconDashboard, IconLogout, IconUpload, IconUser } from '@tabler/icons-react'
+import { AppShell, Button, Group, Stack, Title } from '@mantine/core'
+import { IconAffiliate, IconChevronRight, IconDashboard, IconUpload, IconUser } from '@tabler/icons-react'
 import { useLocation, useNavigate } from 'react-router-dom'
-import { signOut } from 'supertokens-auth-react/recipe/emailpassword'
-import { useGetUsersQuery, useGetOrganizationsQuery } from '@queries'
-import { useSessionContext } from 'supertokens-auth-react/recipe/session'
+import { SignOut } from '../SignOut'
 
 interface ButtonProps {
   name: string
@@ -16,46 +13,6 @@ export const SidePanel = () => {
   const location = useLocation()
   const navigate = useNavigate()
 
-  const [firstName, setFirstName] = useState('User')
-  const [organization, setOrganization] = useState('<Unknown>')
-  const session = useSessionContext()
-  const { data: usersData } = useGetUsersQuery()
-  const { data: organizationsData } = useGetOrganizationsQuery()
-
-  const users = usersData?.users
-  const organizations = organizationsData?.organizations
-
-  useEffect(() => {
-    async function fetchFirstNameAndOrganization() {
-      if (session.loading || !session.doesSessionExist) {
-        return
-      }
-
-      const currentUserID = session.userId
-
-      try {
-        if (currentUserID) {
-          const currentUser = users?.find((user) => user.id === currentUserID)
-          if (currentUser && currentUser.firstName) {
-            setFirstName(currentUser.firstName)
-          }
-
-          if (currentUser && currentUser.organizationId) {
-            const currentOrganization = organizations?.find(
-              (organization) => organization.id === currentUser.organizationId,
-            )
-            if (currentOrganization && currentOrganization.name) {
-              setOrganization(currentOrganization.name)
-            }
-          }
-        }
-      } catch (error) {
-        console.error('Failed to fetch user metadata:', error)
-      }
-    }
-    fetchFirstNameAndOrganization()
-  }, [session, users, organizations])
-
   const buttons: ButtonProps[] = [
     { name: 'Dashboard', Logo: IconDashboard, link: '/' },
     { name: 'Contributions', Logo: IconUpload, link: '/contributions' },
@@ -64,11 +21,6 @@ export const SidePanel = () => {
   ]
 
   const currentPage = buttons.find(({ link }) => link === location.pathname) || buttons[0]
-
-  async function onLogout() {
-    await signOut()
-    navigate('/auth')
-  }
 
   return (
     <>
@@ -95,28 +47,8 @@ export const SidePanel = () => {
           ))}
         </Stack>
       </AppShell.Section>
-      <AppShell.Section style={{ display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
-        <Menu
-          trigger='hover'
-          openDelay={100}
-          closeDelay={400}
-          position='top'
-          offset={8}
-          withArrow
-          arrowPosition='center'
-        >
-          <Menu.Target>
-            <Button variant='outline' color='green' c='gray' size='md' radius='md'>
-              {firstName} from {organization}
-            </Button>
-          </Menu.Target>
-          <Menu.Dropdown>
-            <Menu.Item onClick={onLogout}>
-              <IconLogout size={18} />
-              Sign Out
-            </Menu.Item>
-          </Menu.Dropdown>
-        </Menu>
+      <AppShell.Section style={{ display: 'inline-flex', justifyContent: 'center', alignItems: 'center' }}>
+        <SignOut />
       </AppShell.Section>
     </>
   )
