+++ src/components/index.ts
@@ -14,4 +14,5 @@ export * from './MembersPaper'
 export * from './OrganizationHeader'
 export * from './Paper'
 export * from './SidePanel'
+export * from './SignOut'
 export * from './Theme'
